% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brcutpuw.r, R/netflowpuw.r
\name{brcutpuw}
\alias{brcutpuw}
\alias{netflowpuw}
\title{Linear program based phase unwrapping}
\usage{
brcutpuw(phase, pen = 0, details = FALSE, trace = 1)

netflowpuw(phase, wts = NULL, details = FALSE, trace = 1)
}
\arguments{
\item{phase}{matrix with phase map to be unwrapped (phases are in radians)}

\item{pen}{penalty for making a branch cut from a residue to an edge (\code{brcutpuw} only)}

\item{details}{Return some details of the solution?}

\item{trace}{Send some info from the LP solver to the console?}

\item{wts}{matrix of weights for cost function with same dimension as \code{phase} (\code{netflowpuw} only)}
}
\value{
The unwrapped wavefront in units of fringes.
If \code{details} is \code{TRUE} additional details of the solution
are returned in a named list with the first member
\code{puw} containing the unwrapped wavefront.
}
\description{
\code{brcutpuw} implements a branch cut algorithm for phase unwrapping
}
\details{
This package implements two distinct algorithms for two dimensional
phase unwrapping that can be set up and solved as general linear programs.
The linear programs are solved using the LP solver \code{CBC} from \href{https://github.com/coin-or/Cbc}{COIN-OR}
using the interface provided by the package \href{https://dirkschumacher.github.io/rcbc/index.html}{rcbc}.
}
\section{Functions}{
\itemize{
\item \code{brcutpuw}: Branch cut algorithm for phase unwrapping

\item \code{netflowpuw}: Network flow algorithm for phase unwrapping
}}

\section{Note}{

According to the documentation for \code{rcbc} different levels of detail
from the LP solver can be printed with trace levels up to 15,
however the same output seems to be returned for all values.
Setting \code{trace=0} will produce silent output, which may not
be advisable since these can take some time to run.

The value of trace doesn't matter when running in a Windows "GUI"
console window or Rstudio because the CBC output isn't passed. If
you want to see some output run in a terminal window instead.
}

\examples{
  data("phasemaps", package="lppuw")
  mtext(zernike::rmap(phi, plot=TRUE))
  wf.bc <- brcutpuw(phi)
  wf.nf <- netflowpuw(phi, mod)
  X11()
  zernike::plot.pupil(wf.nf, col=zernike::rygcb(400))
  cat("Summary of the difference between the two unwrapped wavefronts:\n")
  zernike::summary.pupil(wf.nf - wf.bc)

}
\seealso{
There is a function with the same name \code{brcutpuw} in package \link{zernike}.
}
